+++ FOX/recipes/__init__.py
@@ -8,8 +8,10 @@ A number of Auto-FOX recipes.
 
 from .param import get_best, overlay_descriptor, plot_descriptor
 from .psf import generate_psf, generate_psf2, extract_ligand
+from .ligands import get_lig_center
 
 __all__ = [
     'get_best', 'overlay_descriptor', 'plot_descriptor',
-    'generate_psf', 'generate_psf2', 'extract_ligand'
+    'generate_psf', 'generate_psf2', 'extract_ligand',
+    'get_lig_center'
 ]